Update draft-ietf-tsvwg-multipath-dccp.mkd
Replaced use of re-assembly with reorder and reworded slightly to read better.

Also removed the sentence "Such optional reordering mechanisms may account for the fact that packet loss may occur for any of the DCCP subflows" as I think this is incorrect - happy to discuss. Packet loss will cause the reordering mechanism to introduce the maximum delay (based on the length of the reordering buffer), but no reordering happens.